python/requirements.txt
websockets==13.1
faster-whisper==1.1.0
sounddevice==0.5.5
webrtcvad-wheels==2.0.14
requests
huggingface_hub
pyannote.audio